requirements.txt
selenium
pillow
requests
unidecode
tqdm